Extract duplicate InMemoryStringResumeCache to a support class
