+++ components/camel-mongodb/src/test/java/org/apache/camel/processor/resume/mongodb/MongoDbResumeStrategyTest.java
@@ -16,15 +16,9 @@
  */
 package org.apache.camel.processor.resume.mongodb;
 
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Objects;
-import java.util.function.BiFunction;
-import java.util.function.Function;
-
 import org.apache.camel.component.mongodb.MongoDbResumable;
+import org.apache.camel.component.mongodb.support.InMemoryStringResumeCache;
 import org.apache.camel.resume.Cacheable;
-import org.apache.camel.resume.cache.ResumeCache;
 import org.apache.camel.support.processor.state.MemoryStateRepository;
 import org.bson.BsonDocument;
 import org.junit.jupiter.api.Test;
@@ -75,56 +69,4 @@ public class MongoDbResumeStrategyTest {
         assertNull(strategy.getLastOffset("missing"));
     }
 
-    private static final class InMemoryStringResumeCache implements ResumeCache<String> {
-        private final Map<String, Object> data = new HashMap<>();
-
-        @Override
-        public Object computeIfAbsent(String key, Function<? super String, ? super Object> mapping) {
-            return data.computeIfAbsent(key, mapping::apply);
-        }
-
-        @Override
-        public Object computeIfPresent(String key, BiFunction<? super String, ? super Object, ? super Object> remapping) {
-            return data.computeIfPresent(key, remapping::apply);
-        }
-
-        @Override
-        public boolean contains(String key, Object entry) {
-            return Objects.equals(data.get(key), entry);
-        }
-
-        @Override
-        public void add(String key, Object offsetValue) {
-            data.put(key, offsetValue);
-        }
-
-        @Override
-        public boolean isFull() {
-            return false;
-        }
-
-        @Override
-        public long capacity() {
-            return Long.MAX_VALUE;
-        }
-
-        @Override
-        public <T> T get(String key, Class<T> clazz) {
-            return clazz.cast(data.get(key));
-        }
-
-        @Override
-        public Object get(String key) {
-            return data.get(key);
-        }
-
-        @Override
-        public void forEach(BiFunction<? super String, ? super Object, Boolean> action) {
-            for (Map.Entry<String, Object> entry : data.entrySet()) {
-                if (!action.apply(entry.getKey(), entry.getValue())) {
-                    break;
-                }
-            }
-        }
-    }
 }
